+++ CHANGELOG.md
@@ -22,10 +22,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   - Added `Replay Kinesis Messages` doc to Operator Docs.
   - Added `manualConsumer` lambda function to consume a Kinesis stream. Used by the replay AsyncOperation.
 
+- **CUMULUS-1563**
+  - Added `custom_domain_name` variable to `tf-modules/data-persistence` module
+
 - **CUMULUS-1654**
   - Added new helpers to `@cumulus/common/execution-history`:
     - `getStepExitedEvent()` returns the `TaskStateExited` event in a workflow execution history after the given step completion/failure event
     - `getTaskExitedEventOutput()` returns the output message for a `TaskStateExited` event in a workflow execution history
+
+### BREAKING CHANGES
+
+- **CUMULUS-1321**
+  - The `region` argument to the `cumulus` Terraform module has been removed
 
 ### Changed
 
@@ -43,18 +51,30 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - **CUMULUS-1654**
   - Changed `@cumulus/common/SfnStep.parseStepMessage()` to a static class method
 
+- **CUMULUS-1641**
+  - Added `meta.retries` and `meta.visibilityTimeout` properties to sqs-type rule. To create sqs-type rule, you're required to configure a dead-letter queue on your queue.
+  - Added `sqsMessageRemover` lambda which removes the message from SQS queue upon successful workflow execution.
+  - Updated `sqsMessageConsumer` lambda to not delete message from SQS queue, and to retry the SQS message for configured number of times.
+
 ### Removed
 
 - Removed `create_service_linked_role` variable from `tf-modules/data-persistence` module.
 
 ### Fixed
 
+- **CUMULUS-1668** - Fixed a race condition where executions may not have been
+  added to the database correctly
 - **CUMULUS-1654** - Fixed issue with `publishReports` Lambda not including workflow execution error information for failed workflows with a single step
 - Fixed `tf-modules/cumulus` module so that the `urs_url` variable is passed on to its invocation of the `tf-modules/archive` module
 
 ## [v1.16.0] - 2019-11-15
 
 ### Added
+
+- **CUMULUS-1321**
+  - A `deploy_distribution_s3_credentials_endpoint` variable has been added to
+    the `cumulus` Terraform module. If true, the NGAP-backed S3 credentials
+    endpoint will be added to the Thin Egress App's API. Default: true
 
 - **CUMULUS-1544**
   - Updated the `/granules/bulk` endpoint to correctly query Elasticsearch when